requirements.txt
requests>=2.31.0
pandas>=2.0.0
duckdb>=0.9.0
prefect>=2.14.0
pyyaml>=6.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
tqdm>=4.65.0
pyarrow>=12.0.0
nba-api>=1.2.1
